Quand le secret est invalide, aller à la page d'accueil avec un message d'erreur
